finish RWD of our team section
